- Fixed warnings in main.cpp and keyboard.cpp - Changed Vector2 and Matrix3x3 classes to be generic based, which removed the silly MathType type. - Minor change to path finding code so that instead of inserting into the front of the final path, it simply reverses the result which is much faster. - Added code for finding the closest of a given entity type based on path distance to the entity from a starting point. - Cleaned up the Game menu code a bit, so that it's more menu like to support more in depth menus.

+++ simplerpg/UIContainer.cpp
@@ -52,7 +52,7 @@ void UIContainer::removeChild(UIComponent &child)
 
 void UIContainer::removeChild(int index)
 {
-	if(index >= 0 && index < mChildren.size())
+	if(index >= 0 && index < (int)mChildren.size())
 	{
 		removeChild(*mChildren[index]);
 	}
